pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unemployment-tracker"
version = "0.1.0"
description = "A comprehensive tool for tracking and analyzing unemployment data"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "requests>=2.26.0",
    "python-dotenv>=0.19.0",
    "streamlit>=1.10.0",
    "plotly>=5.5.0",
    "geopandas>=0.10.0",
    "folium>=0.12.0",
    "scikit-learn>=1.0.0",
    "networkx>=2.6.0",
    "pytest>=6.2.5",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.930",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=3.0.0",
    "pylint>=2.12.0",
    "black[jupyter]>=22.0.0",
    "pre-commit>=2.17.0",
]

[project.scripts]
unemployment-tracker = "unemployment_tracker.app:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
